requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pandas
numpy
joblib
sentence-transformers
faiss-cpu
pypdf
requests
ibm-watsonx-ai
